+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.OurMenu"
         tools:targetApi="31">
         <activity
-            android:name=".newfolder.PostMenuFolderActivity"
+            android:name=".menu.menufolder.PostMenuFolderActivity"
             android:exported="false" />
         <activity
             android:name=".menu.menufolder.MenuFolderActivity"
