+++ app/src/main/java/com/example/ourmenu/menu/menufolder/PostMenuFolderGetDetailFragment.kt
@@ -1,11 +1,10 @@
-package com.example.ourmenu.newfolder
+package com.example.ourmenu.menu.menufolder
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.ourmenu.R
 import com.example.ourmenu.databinding.FragmentPostMenuFolderGetDetailBinding
 
 class PostMenuFolderGetDetailFragment : Fragment() {
