+++ app/src/main/java/com/example/ourmenu/menu/menufolder/PostMenuFolderGetFragment.kt
@@ -1,22 +1,21 @@
-package com.example.ourmenu.newfolder
+package com.example.ourmenu.menu.menufolder
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.ourmenu.databinding.FragmentPostMenuFolderBinding
+import com.example.ourmenu.databinding.FragmentPostMenuFolderGetBinding
 
-class PostMenuFolderFragment : Fragment() {
+class PostMenuFolderGetFragment : Fragment() {
 
-    lateinit var binding: FragmentPostMenuFolderBinding
+    lateinit var binding: FragmentPostMenuFolderGetBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentPostMenuFolderBinding.inflate(layoutInflater)
-
+        binding = FragmentPostMenuFolderGetBinding.inflate(layoutInflater)
