+++ app/src/main/res/layout/activity_post_menu_folder.xml
@@ -5,7 +5,7 @@
     android:id="@+id/post_menu_folder_frm"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".newfolder.PostMenuFolderActivity">
+    tools:context=".menu.menufolder.PostMenuFolderActivity">
 
 
 </FrameLayout>
