+++ app/src/main/res/layout/fragment_post_menu_folder_get.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".newfolder.PostMenuFolderGetFragment">
+    tools:context=".menu.menufolder.PostMenuFolderGetFragment">
 
 
     <ImageView
