+++ app/src/main/res/layout/fragment_post_menu_folder_get_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".newfolder.PostMenuFolderGetDetailFragment">
+    tools:context=".menu.menufolder.PostMenuFolderGetDetailFragment">
 
 
     <ImageView
